refactor(in-app-agent): serialize the sweeps and drop redundant work
Two concurrent sweeps were possible. BullMQ gives one delivery per interval,
not one execution at a time: worker.js produces the next scheduled job when
the current one is picked up for processing, so a tick that outruns its 5s
interval overlaps with the next tick on another worker task. Every operation
is a guarded CAS and survives that, but overlap only happens when a tick is
slow, which is exactly when duplicating its scans is worst. Take a Redis lease
around each sweep, failing closed if Redis is unavailable rather than fanning
out. Cadence still comes from BullMQ, so this does not reintroduce the
per-worker timer multiplier the RedisLock-only design had.

Then cut what stopped paying for itself:

- Drop the terminal-pointer credential batch. Now that terminalization revokes
  the credential inline, it only covered a run whose own delete failed, and
  the orphan backstop already finds those keys by age and clears their
  pointers. Removes the second unindexed sequential scan.
- Derive the oldest-queued-run gauge from the candidate rows already read
  instead of a second query per tick. Runs younger than the redispatch delay
  are invisible to it, which is fine: the canary monitor fires at 30s.
- Merge the two worker modules into lifecycleSweeps.ts. With the batch gone,
  credential work is one function, and the split had lifecycleRecovery
  importing from credentialMaintenance to revoke inline.
- Drop the unused createdAt on redispatch entries.

Production diff drops from ~845 to ~760 lines, tests from ~545 to ~466, and a
healthy tick now costs one query instead of two.

Refs LFE-14624
